refactor: extract events to data and router
